Allow workdir be computed dynamically based on sourcestamp.
The patch allows one to set workdir on BuildFactory to a
callable which receives the sourcestamp to be built and is
supposed to return a workdir directory.

A reason to do so is having multiple repositories submitting
changes to a single buildbot, where one wants to allow for
workdir per repository.

You can achieve that like so:

import hashlib

build_factory = factory.BuildFactory()
build_factory.workdir = lambda source_stamp: hashlib.md5(source_stamp.repository).hexdigest()[:8]