feature: dynamic buildset using builderNames property of change
The patch allows one to control the exact set of builders at change
should be built on using the properties bag of a change.

This works with Scheduler/AnyBranchScheduler.

When a change contains something like

  properties['builderNames'] = ["builder1", "builder2"]

and the change triggers a scheduler, the scheduler will create
a buildset which targets the given builders.

When a change does not contain such a property, the builder list
of the scheduler defined will apply.

Sanity checks are made (property must be list/dict, builder names
must be strings, and: builder names must be contained in the list
of builder provided to the constructor of Scheduler!).